Clean up and add unit tests for AdAdaptor.java
As part of the testing framework, imported a few helper test files
(AccumulatingDocIdPusher, DelegatingDocIdPusher, and UnsupportedDocIdPusher)
from the Sharepoint library, as well as TestHelper from the standard library.
(Only a couple of methods - those actually used - were retained there.)

This submission also removes the AdConstants and AdMembership classes, as
they are no longer needed.  In addition, it adds an equal() method to both
the AdAdaptor.GroupProperties and AdEntity classes.