Patch unit tests for HashSet ordering in Java 8.
This is a hack for treating the List<Principal> in the groups map as
unordered. This is a small change because an assertEquals was expanded
to the java.util.AbstractMap#equals implementation in commit 592b00e.

Two more intrusive fixes are to change the internal
Map<GroupPrincipal, List<Principal>> to use a Set instead of a Map, so
the comparison is always unordered, or to change the members map in
AdEntity to a TreeMap, for predictable ordering (using LinkedHashMap
does not work, though that may be due to the setup in the tests). The
performance of those changes would need to be tested at scale, and
it's only the tests that care about the ordering.